Task #4956: undoing previous merges

+++ branches/LOFAR-preRelease-1_17/RTCP/Cobalt/GPUProc/src/cuda/Kernels/DelayAndBandPassKernel.h
@@ -88,7 +88,7 @@ namespace LOFAR
 
 
       void enqueue(PerformanceCounter &counter,
-                   double subbandFrequency, size_t SAP);
+                   float subbandFrequency, size_t SAP);
     };
 
     // Specialization of the KernelFactory for
